#1313 Rename DomesticTax,ForeignTax and review fixes
